+++ tests/api/schema/test_schema_action.py
@@ -16,7 +16,6 @@ def test_api_schema_action_defaults():
     assert not action.is_collection()
     assert action.get_entity_path() == ''
     assert action.get_path_delimiter() == '/'
-    assert action.get_primary_key() == 'id'
     assert action.resolve_entity({}) == {}
     assert not action.has_entity_definition()
     assert action.get_entity() == {}
@@ -71,7 +70,6 @@ def test_api_schema_action(read_json):
     assert action.is_collection()
     assert action.get_entity_path() == payload.get('entity_path')
     assert action.get_path_delimiter() == payload.get('path_delimiter')
-    assert action.get_primary_key() == payload.get('entity/primary_key')
     assert action.resolve_entity({'foo': {'bar': 'OK'}}) == 'OK'
     # Resolve an invalid entity
     with pytest.raises(ActionSchemaError):